chore: Forge refactor: Extract forge into a separate file/module
