+++ ouroboros-consensus.cabal
@@ -1113,6 +1113,7 @@ library diffusion
     Ouroboros.Consensus.Node.RethrowPolicy
     Ouroboros.Consensus.Node.Tracers
     Ouroboros.Consensus.NodeKernel
+    Ouroboros.Consensus.NodeKernel.Forge
 
   reexported-modules:
     Ouroboros.Consensus.Block,
